Lots of changes, a lot of products added. Added slider on product page, qty is now limited to avbl amount on product page. Filters for flavour are multi-choice. Admin checks are added. more easy to use admin panel.

+++ php/admin_delete.php
@@ -1,10 +1,12 @@
 <?php
 require_once './db.php';
+require_once 'is_admin.php';
+require_admin();
 
 $data = json_decode(file_get_contents("php://input"), true);
 $table = $data['table'] ?? '';
 $id = $data['id'] ?? '';
-
+echo $id;
 if ($table && $id) {
     $stmt = $pdo->prepare("DELETE FROM `$table` WHERE id = :id");
     $stmt->execute(['id' => $id]);
